fetch stories chronologically
My goal with this is to fetch stories completely chronologically, not favoring any based on their rank. The old code used the new/top post API, but this API only returned 500 posts per category. For ML like this, more than 500 articles might be needed.

It should be noted that, as part of the listing process, I ensure that the discovered articles are from a few days ago, so that scores have "settled".