normalize feature vectors
Before now, the neuralnet was quickly converging to a state where it would always output the same thing for every input. As a result, it would always predict the most frequent class of article. This actually got it 42/88 on the same data I tried with, which seems impressive but really is not.